load_testing/requirements.txt
# Load Testing Requirements
# Additional packages needed for comprehensive load testing

# Core load testing framework
locust>=2.17.0

# System monitoring
psutil>=5.9.0

# HTTP requests
requests>=2.31.0

# Image processing for synthetic test data
Pillow>=10.0.0

# Scientific computing for test data generation
numpy>=1.24.0

# Data analysis for results
pandas>=2.0.0

# Visualization for dashboard
plotly>=5.15.0

# Statistics and math
scipy>=1.11.0
